PR fiks
